// src/app/core/enums/servicos.ts
export enum Servicos {
    faltaDeEnergia = 'falta-de-energia',
    religacao = 'religacao',
    historicoDeconsumo = 'historico-de-consumo',
    autoleitura = 'autoleitura',
    debitoAutomatico = 'debito-automatico',
    dataCerta = 'data-certa',
    lgpd = 'lgpd',
    minhaConta = 'minha-conta',
    segundaVia = 'consultar-debitos',
    alterarDataDeVencimento = 'data-certa',
    acompanheSeuPedido = 'acompanhe-seu-pedido',
    escolhaTipoFatura = 'escolha-tipo-fatura',
    faturaImpressa = 'fatura-impressa',
    faturaDigital = 'fatura-digital',
    faturaBraile = 'fatura-em-braile',
    entendaSuaConta = 'entenda-sua-conta',
    trocaDeTitularidade = 'troca-de-titularidade',
    memoriaDeMassa = 'memoria-de-massa',
    solicitacaoEnviada = 'solicitacao-enviada',
    segundaViaLogin = 'segunda-via-de-pagamento',
    cadastro = 'cadastro',
    aviso = 'aviso',
    multiloginAcesso = 'multilogin-acesso',
    multiloginCadastroCredenciado = 'multilogin-cadastro-credenciado',
    multiloginCadastroImobiliaria = 'multilogin-cadastro-imobiliario',
    multiloginCadastro = 'multilogin-cadastro',
    recuperarSenha = 'recuperar-senha',
    projetoParticular = 'projeto-particular',
    projetoParticularSolicitarAnalise = 'solicitar-analise',
    projetoParticularFormulario = 'formulario',
    unidadesConsumidoras = 'meus-imoveis',
    compartilharAcesso = 'compartilhar-acesso',
    meusImoveis = 'meus-imoveis',
    ligacaoNova = 'ligacao-nova',
    selecionarPerfilAcesso = 'selecionar-perfil-de-acesso'
}

export enum PathCompleto {
    paginaInicial = '/login' + '/',
    home = '/home' + '/',
    meusImoveis = '/home/' + Servicos.meusImoveis,
    faltaDeEnergia = '/home/servicos/' + Servicos.faltaDeEnergia,
    religacao = '/home/servicos/' + Servicos.religacao,
    historicoDeconsumo = '/home/servicos/' + Servicos.historicoDeconsumo,
    autoleitura = '/home/servicos/' + Servicos.autoleitura,
    debitoAutomatico = '/home/servicos/' + Servicos.debitoAutomatico,
    dataCerta = '/home/servicos/' + Servicos.dataCerta,
    minhaConta = '/home/' + Servicos.minhaConta,
    segundaVia = '/home/servicos/' + Servicos.segundaVia,
    alterarDataDeVencimento = '/home/servicos/' + Servicos.alterarDataDeVencimento,
    acompanheSeuPedido = '/home/servicos/' + Servicos.acompanheSeuPedido,
    escolhaTipoFatura = '/home/servicos/' + Servicos.escolhaTipoFatura,
    faturaImpressa = '/home/servicos/' + Servicos.faturaImpressa,
    faturaDigital = '/home/servicos/' + Servicos.faturaDigital,
    faturaBraile = '/home/servicos/' + Servicos.faturaBraile,
    entendaSuaConta = '/home/servicos/' + Servicos.entendaSuaConta,
    trocaDeTitularidade = '/home/servicos/' + Servicos.trocaDeTitularidade,
    memoriaDeMassa = '/home/servicos/' + Servicos.memoriaDeMassa,
    segundaViaLogin = '/login/' + Servicos.segundaViaLogin,
    cadastro = '/login/' + Servicos.cadastro,
    aviso = '/home/servicos/' + Servicos.aviso,
    multiloginAcesso = '/home/' + Servicos.multiloginAcesso,
    multiloginCadastroCredenciado = '/login/' + Servicos.multiloginCadastroCredenciado,
    multiloginCadastroImobiliario = '/login/' + Servicos.multiloginCadastroImobiliaria,
    multiloginCadastro = '/home/' + Servicos.multiloginCadastro,
    recuperarSenha = '/login/' + Servicos.recuperarSenha,
    unidadesConsumidoras = '/home/' + Servicos.unidadesConsumidoras,
    compartilharAcesso = '/home/' + Servicos.compartilharAcesso,
    projetoParticular = '/home/servicos/' + Servicos.projetoParticular,
    ligacaoNova = '/' + Servicos.ligacaoNova,
    selecionarPerfilAcesso = '/home/' + Servicos.multiloginAcesso + '/' + Servicos.selecionarPerfilAcesso
}